+++ src/pptx/oxml/__init__.py
@@ -247,6 +247,8 @@ register_element_cls("c:tx", CT_Tx)
 register_element_cls("c:varyColors", CT_Boolean)
 register_element_cls("c:x", CT_Double)
 register_element_cls("c:xMode", CT_LayoutMode)
+register_element_cls("c:y", CT_Double)
+register_element_cls("c:yMode", CT_LayoutMode)
 
 
 from pptx.oxml.comments import (  # noqa: E402